:truck: folder and files renamed
index.ts renamed to cli.ts and bot folder renamed to lib
